+++ miamoremustachio/ToDo/modules/routes/task.js
@@ -1,0 +1,94 @@
+const express = require('express');
+
+const { Task } = require('../models/task.js');
+const { User } = require('../models/user.js');
+const { findTask } = require('../middlewares/task_searching.js');
+const { checkUserId } = require('../middlewares/authorization.js');
+const { Query } = require('../helpers/constructors.js');
+const {
+  checkTitle,
+  checkProperties,
+} = require('../helpers/checking.js');
+
+const router = express.Router();
+
+router.route('/')
+  .get(async (req, res) => {
+    const userId = req.get('Authorization');
+    const query = { userId, ...req.query };
+    
+    try {      
+      const tasksList = await Task.find(query);
+  
+      res.json(tasksList);
+    } catch(error) {
+      res.status(400).send(error.message);
+    }
+  })
+  .post(async (req, res) => {
+    const userId = req.get('Authorization');
+    const { title, ...restProperties } = req.body;
+
+    try {
+      checkTitle(title);
+      checkProperties(restProperties);
+
+      const task = new Task({ userId, ...req.body });
+
+      await task.save();
+
+      const taskPath = `${req.baseUrl}${req.path}${task.id}`;
+
+      res.status(201).send(taskPath);
+    } catch(error) {
+      res.status(400).send(error.message);
+    }
+  });
+
+  router.route('/:id')
+  .all(findTask, checkUserId)
+  .get(async (req, res) => {
+    const taskId = req.params.id;
+    const userId = req.get('Authorization');
+
+    try {
+      const task = await Task.findById(taskId);
+      const user = await User.findById(userId);
+
+      const taskObject = task.toObject();
+      taskObject.user = user;
+      delete taskObject.userId;
+
+      res.json(taskObject);
+    } catch(error) {
+      res.status(500).send(error.message);
+    }
+  })
+  .put(async (req, res) => {
+    const taskId = req.params.id;
+
+    try {
+      checkProperties(req.body);
+
+      const query = new Query(req.body);
+      const options = { returnDocument: "after" };
+      const result = await Task.findByIdAndUpdate(taskId, query, options);
+
+      res.send(result);
+    } catch(error) {
+      res.status(400).send(error.message);
+    }
+  })
+  .delete(async (req, res) => {
+    const taskId = req.params.id;
+
+    try {
+      const result = await Task.findByIdAndDelete(taskId);
+      
+      res.send(result);
+    } catch(error) {
+      res.status(500).send(error.message);
+    }
+  })
+
+module.exports = { router };
